refactor: authentication blueprint
