+++ microblog/app/__init__.py
@@ -31,6 +31,9 @@ babel = Babel(app, locale_selector=get_locale)
 from app.errors import bp as errors_bp
 app.register_blueprint(errors_bp)
 
+from app.auth import bp as auth_bp
+app.register_blueprint(auth_bp, url_prefix='/auth')
+
 if not app.debug:
     if app.config['MAIL_SERVER']:
         auth = None
