+++ microblog/app/email.py
@@ -1,4 +1,3 @@
-from flask import render_template
 from flask_mail import Message
 from threading import Thread
 from app import app, mail
@@ -12,11 +11,3 @@ def send_email(subject, sender, recipients, text_body, html_body):
     msg.body = text_body
     msg.html = html_body
     Thread(target=send_async_email, args=(app, msg)).start()
-
-def send_password_reset_email(user):
-    token = user.get_reset_password_token()
-    send_email('[Microblog] Reset Your Password',
-              sender=app.config['ADMINS'][0],
-              recipients=[user.email],
-              text_body=render_template('email/reset_password.txt',user=user, token=token),
-              html_body=render_template('email/reset_password.html', user=user, token=token))
